% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{boxplot_runoff_volume}
\alias{boxplot_runoff_volume}
\title{Boxplot Runoff Volume per Event}
\usage{
boxplot_runoff_volume(
  lid = "bioretention_cell",
  zone_id = 5,
  performances = keys.lid::performances
)
}
\arguments{
\item{lid}{tidy name of LID (default: "bioretention_cell")}

\item{zone_id}{climate zone id to plot (default: 1)}

\item{performances}{nested tibble (default: \code{\link{performances}})}
}
\value{
interactive plot of performance results
}
\description{
Boxplot Runoff Volume per Event
}
\examples{
\dontrun{
boxplot_runoff_volume(lid = "bioretention_cell", zone_id = 1)
}
}
