% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{boxplot_vrr}
\alias{boxplot_vrr}
\title{Boxplot Volume Rainfall Retended per Year}
\usage{
boxplot_vrr(
  lid = "bioretention_cell",
  zone_id = 1,
  performances = keys.lid::performances
)
}
\arguments{
\item{lid}{tidy name of LID (default: "bioretention_cell")}

\item{zone_id}{climate zone id to plot (default: 1)}

\item{performances}{nested tibble (default: \code{\link{performances}})}
}
\value{
interactive plot of performance results
}
\description{
Boxplot Volume Rainfall Retended per Year
}
\examples{
\dontrun{
boxplot_vrr(lid = "bioretention_cell", zone_id = 1)
}
}
