% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_roberto.R
\name{computeVol}
\alias{computeVol}
\title{computeVol}
\usage{
computeVol(data, timeColumn, Qcolumn)
}
\arguments{
\item{data}{data}

\item{timeColumn}{timeColumn}

\item{Qcolumn}{Qcolumn}
}
\value{
???
}
\description{
compute runoff volume for runoff in mm/s
}
