% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_performances.R
\name{export_performances}
\alias{export_performances}
\title{Title}
\usage{
export_performances(export_dir = tempdir())
}
\arguments{
\item{export_dir}{default: tempdir()}
}
\value{
write "performances" to "swmm_lid-performances.xlsx" in directory
"export_dir" and return path to fike
}
\description{
Title
}
