% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extdata_file}
\alias{extdata_file}
\title{Get Path to File in This Package}
\usage{
extdata_file(...)
}
\arguments{
\item{\dots}{parts of path passed to \code{\link{system.file}}}
}
\description{
Get Path to File in This Package
}
