% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_percentiles.R
\name{get_event_percentiles}
\alias{get_event_percentiles}
\title{Get Percentiles for Events}
\usage{
get_event_percentiles(performances = keys.lid::performances)
}
\arguments{
\item{performances}{nested tibble (default: \code{\link{performances}})}
}
\value{
list with percentiles for "event_sum" and "event_max"
}
\description{
Get Percentiles for Events
}
