% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidconfig_to_swmm.R
\name{lidconfig_to_swmm}
\alias{lidconfig_to_swmm}
\title{Convert LID config to SWMM LID controls}
\usage{
lidconfig_to_swmm(df)
}
\arguments{
\item{df}{data frame for a single scenario of a LID (as returned by
\code{\link{read_scenarios}})}
}
\value{
data frame with SWMM LID controls
}
\description{
Convert LID config to SWMM LID controls
}
\examples{
scenarios <- keys.lid::read_scenarios()
unique(scenarios$lid_name_tidy)
lid <- "permeable_pavement"
lid_selected <- scenarios \%>\%  dplyr::filter(.data$lid_name_tidy == lid)
scenario_names <- unique(lid_selected$scenario_name)
scenario_name <- scenario_names[1]
scenario_name
lid_selected_scenario <- lid_selected[lid_selected$scenario_name == scenario_name,]
lid_controls <- lidconfig_to_swmm(lid_selected_scenario)
str(lid_controls)
}
