% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_roberto.R
\name{makeRainfallRunoffEvents}
\alias{makeRainfallRunoffEvents}
\title{makeRainfallRunoffEvents}
\usage{
makeRainfallRunoffEvents(rainfalldata, runoffdata)
}
\arguments{
\item{rainfalldata}{rainfalldata}

\item{runoffdata}{runoffdata}
}
\value{
????
}
\description{
makeRainfallRunoffEvents
}
