% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vrr_median.R
\name{plot_vrr_median}
\alias{plot_vrr_median}
\title{Plot Median VRR}
\usage{
plot_vrr_median(lid = "bioretention_cell")
}
\arguments{
\item{lid}{tidy name of LID}
}
\value{
interactive plot of performance results
}
\description{
Plot Median VRR
}
\examples{
\dontrun{
lids <- unique(keys.lid::performances$lid_name_tidy)
sapply(lids, function(lid) print(keys.lid::plot_vrr_median(lid)))
}
}
