% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scenarios.R
\name{read_scenarios}
\alias{read_scenarios}
\title{read_scenarios}
\usage{
read_scenarios(
  scenarios_xlsx = extdata_file("scenarios/swmm_lid-parameterisation.xlsx")
)
}
\arguments{
\item{scenarios_xlsx}{path to LID scenarios Excel file (default:
\code{\link{extdata_file}} with: "scenarios/swmm_lid-parameterisation.xlsx"}
}
\value{
tidy scenarios data frame
}
\description{
read_scenarios
}
