% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_roberto.R
\name{readObservations}
\alias{readObservations}
\title{readObservations}
\usage{
readObservations(
  subfolder,
  rainFile,
  runoffFile,
  temperatureFile,
  dateTimetz,
  dateTimeformat,
  to_mmperhour,
  NAval
)
}
\arguments{
\item{subfolder}{subfolder}

\item{rainFile}{rainFile}

\item{runoffFile}{runoffFile}

\item{temperatureFile}{temperatureFile}

\item{dateTimetz}{dateTimetz}

\item{dateTimeformat}{dateTimeformat}

\item{to_mmperhour}{to_mmperhour}

\item{NAval}{NAval}
}
\value{
????
}
\description{
readObservations
}
