% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_roberto.R
\name{readPredictions}
\alias{readPredictions}
\title{readPredictions}
\usage{
readPredictions(
  subfolder,
  rainFile,
  runoffFile,
  temperatureFile,
  dateTimetz,
  dateTimeformat,
  to_mmperhour,
  parTcontinuous
)
}
\arguments{
\item{subfolder}{subfolder}

\item{rainFile}{rainFile}

\item{runoffFile}{runoffFile}

\item{temperatureFile}{temperatureFile}

\item{dateTimetz}{dateTimetz}

\item{dateTimeformat}{dateTimeformat}

\item{to_mmperhour}{to_mmperhour}

\item{parTcontinuous}{parTcontinuous}
}
\value{
????
}
\description{
readPredictions
}
