% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_performance.R
\name{simulate_performance}
\alias{simulate_performance}
\title{Simulate Performance of LID}
\usage{
simulate_performance(
  lid_selected,
  lid_area_fraction = 0,
  catchment_area_m2 = 1000,
  col_eventsep = "total_rainfall",
  swmm_base_inp = keys.lid::extdata_file("scenarios/models/model_template.inp"),
  swmm_climate_dir = keys.lid::extdata_file("rawdata/weather_sponge_regions"),
  model_dir = keys.lid::extdata_file("scenarios/models"),
  zone_ids = 1L:5L
)
}
\arguments{
\item{lid_selected}{tibble with a selected LID as retrieved by \code{\link{read_scenarios}}}

\item{lid_area_fraction}{fraction of LID in subcatchment (default: 0)}

\item{catchment_area_m2}{catchment area (default: 1000 m2)}

\item{col_eventsep}{SWMM output column used for event separation (default:
"total_rainfall")}

\item{swmm_base_inp}{path to SWMM model to be used as template for modification
(default: keys.lid::extdata_file("scenarios/models/model_template.inp"))}

\item{swmm_climate_dir}{directory with climate data
(default: keys.lid::extdata_file("rawdata/weather_sponge_regions")}

\item{model_dir}{default:  keys.lid::extdata_file("scenarios/models")}

\item{zone_ids}{climate zone ids to be used for simulation (default: 1L:5L)}
}
\value{
tibble with nested lists containing all scenario performance
}
\description{
Simulate Performance of LID
}
\examples{
\dontrun{
scenarios <- keys.lid::read_scenarios()
unique(scenarios$lid_name_tidy)
lid <- "permeable_pavement"
lid_selected <- scenarios \%>\%  dplyr::filter(.data$lid_name_tidy == lid)
pp_0.00 <- keys.lid::simulate_performance(lid_selected,
                                          lid_area_fraction = 0.00)
pp_1.0 <- keys.lid::simulate_performance(lid_selected,
                                         lid_area_fraction = 1.0)
pp <- dplyr::bind_rows(pp_0.00, pp_1.0)
}
}
