% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squaremeter_to_hectar.R
\name{squaremeter_to_hectar}
\alias{squaremeter_to_hectar}
\title{Helper function: squaremeter to hectr}
\usage{
squaremeter_to_hectar(squaremeter)
}
\arguments{
\item{squaremeter}{squarmeter}
}
\value{
area in hectar
}
\description{
Helper function: squaremeter to hectr
}
\examples{
squaremeter_to_hectar(100)
}
